Creation of a patch that allow the slime to dash but it doesnt work on a ramp
The problem is that with the move and collide the slime collide with the ground and stop it movement.
The function adjust height should stop the slime to do that but it does'nt work like planed with the new collider that I had (this collider is more precise but I didn't ajust it like I should and I don't know how to do it ).
Need to check all of that more precisly in order to fix this problem